Initial relocation of server types
