+++ .codesandbox/ci.json
@@ -3,8 +3,7 @@
   "installCommand": "install-with-npm-8.5",
   "packages": [
     "packages/usage-reporting-protobuf",
-    "packages/server",
-    "packages/server-types"
+    "packages/server"
   ],
   "sandboxes": ["apollo-server-typescript-3opde","apollo-server"],
   "node": "16"
