--- packages/server-types/.npmignore
@@ -1,7 +1,0 @@
-*
-!src/**/*
-src/**/__tests__/**
-!dist/**/*
-dist/**/__tests__/**
-!package.json
-!README.md
