--- packages/server-types/README.md
@@ -1,4 +1,0 @@
-# `apollo-server-types`
-
-These are types which are shared across Apollo Server packages, but kept here
-to avoid circular dependencies.
