--- packages/server-types/jest.config.js
@@ -1,3 +1,0 @@
-const config = require('../../jest.config.base');
-
-module.exports = Object.assign(Object.create(null), config);
