--- packages/server-types/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@apollo/server-types",
-  "version": "3.5.1",
-  "description": "Apollo Server shared types",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "keywords": [],
-  "author": "Apollo <packages@apollographql.com>",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/apollographql/apollo-server",
-    "directory": "packages/server-types"
-  },
-  "engines": {
-    "node": ">=14.0"
-  },
-  "dependencies": {
-    "@apollo/usage-reporting-protobuf": "^3.3.0",
-    "apollo-server-caching": "^3.3.0",
-    "@types/node-fetch": "^2.5.12"
-  },
-  "peerDependencies": {
-    "graphql": "^16.3.0"
-  },
-  "volta": {
-    "extends": "../../package.json"
-  }
-}
