--- packages/server-types/src/index.ts
@@ -1,10 +1,0 @@
-// Temporary index file for AS4 `@apollo/server-types`. This package will be
-// renamed to `@apollo/server-types` and the API will be checked to ensure it
-// contains only types needed to build plugins and other extensions (not
-// internal functionality of AS).
-
-// These are the types that used to live in `apollo-server-plugin-base`.
-export * from './plugins';
-
-// These are the types that used to live in `@apollo/server-types`.
-export * from './types';
