+++ packages/server/package.json
@@ -25,7 +25,6 @@
     "node": ">=12.0"
   },
   "dependencies": {
-    "@apollo/server-types": "^3.5.1",
     "@apollo/usage-reporting-protobuf": "^3.3.0",
     "@apollo/utils.logger": "^1.0.0",
     "@apollographql/graphql-playground-html": "1.6.29",
