+++ packages/server/src/ApolloServer.ts
@@ -32,7 +32,7 @@ import type {
   HTTPGraphQLRequest,
   HTTPGraphQLResponse,
   WithRequired,
-} from '@apollo/server-types';
+} from './types/index';
 import type { Logger } from '@apollo/utils.logger';
 
 import type {
