+++ packages/server/src/__tests__/ApolloServer.test.ts
@@ -1,7 +1,7 @@
 import { ApolloServer } from '../ApolloServer';
 import { ApolloServerOptions, GatewayInterface, gql } from '..';
 import type { GraphQLSchema } from 'graphql';
-import type { ApolloServerPlugin, BaseContext } from '@apollo/server-types';
+import type { ApolloServerPlugin, BaseContext } from '../types/index';
 import {
   ApolloServerPluginCacheControlDisabled,
   ApolloServerPluginUsageReporting,
