+++ packages/server/src/__tests__/cachePolicy.test.ts
@@ -1,4 +1,4 @@
-import type { CachePolicy } from '@apollo/server-types';
+import type { CachePolicy } from '../types/index';
 import { newCachePolicy } from '../cachePolicy';
 
 describe('newCachePolicy', () => {
