+++ packages/server/src/__tests__/tsconfig.json
@@ -8,7 +8,6 @@
     "lib": ["es2019", "es2019.array", "esnext.asynciterable", "dom"],
   },
   "references": [
-    { "path": "../../" },
-    { "path": "../../../server-types" },
+    { "path": "../../" }
   ]
 }
