+++ packages/server/src/internalPlugin.ts
@@ -1,4 +1,4 @@
-import type { BaseContext, ApolloServerPlugin } from '@apollo/server-types';
+import type { BaseContext, ApolloServerPlugin } from './types/index';
 
 // This file's exports should not be exported from the overall
 // @apollo/server module.
