+++ packages/server/src/plugin/landingPage/default/index.ts
@@ -1,4 +1,4 @@
-import type { BaseContext } from '@apollo/server-types';
+import type { BaseContext } from '../../../types/index';
 import type { ImplicitlyInstallablePlugin } from '../../../ApolloServer';
 
 export interface ApolloServerPluginLandingPageDefaultBaseOptions {
