+++ packages/server/src/plugin/usageReporting/options.ts
@@ -4,7 +4,7 @@ import type {
   GraphQLRequestContext,
   GraphQLRequestContextWillSendResponse,
   BaseContext,
-} from '@apollo/server-types';
+} from '../../types/index';
 import type { Logger } from '@apollo/utils.logger';
 import type fetch from 'node-fetch';
 import type { Agent as HttpAgent } from 'http';
