+++ packages/server/src/runHttpQuery.ts
@@ -4,7 +4,7 @@ import type {
   BaseContext,
   HTTPGraphQLRequest,
   HTTPGraphQLResponse,
-} from '@apollo/server-types';
+} from './types/index';
 import {
   ApolloServerInternals,
   internalExecuteOperation,
