+++ packages/server/src/types.ts
@@ -16,7 +16,7 @@ import type {
   BaseContext,
   GraphQLResponse,
   GraphQLRequestContext,
-} from '@apollo/server-types';
+} from './types/index';
 import type { Logger } from '@apollo/utils.logger';
 
 import type { KeyValueCache } from 'apollo-server-caching';
