+++ packages/server/src/types/index.ts
@@ -1,0 +1,6 @@
+// FIXME: this goes away once types all find their proper homes
+// These are the types that used to live in `apollo-server-plugin-base`.
+export * from './plugins';
+
+// These are the types that used to live in `@apollo/server-types`.
+export * from './types';
