+++ packages/server/src/types/plugins.ts
