+++ tsconfig.build.json
@@ -5,7 +5,6 @@
   "files": [],
   "include": [],
   "references": [
-    { "path": "./packages/server" },
-    { "path": "./packages/server-types" },
+    { "path": "./packages/server" }
   ]
 }
